requirements.txt
python-dotenv==0.20.0
aiogram==2.20
SQLAlchemy==1.4.35